changed the widget name to UserBookings
