--- lib/widget/Bookings/bookingGestureDetector.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class BookingGestureDetector extends StatelessWidget {
-  const BookingGestureDetector({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return const Placeholder();
-  }
-}
